fix(workspace): address review — move type to lib/types, stable callback, add test
- Move WorkspaceStatusInfo to lib/types.ts (domain type belongs with other types)
- Extract onBackToDashboard as stable useCallback prop in App.tsx
- Add test for missing statusInfo on active workspace
- Replace `as CiStatus` casts with `satisfies` in test fixtures
