+++ src/components/Workspace/index.tsx
@@ -1,2 +1,2 @@
 export { WorkspaceView } from "./WorkspaceView";
-export type { WorkspaceStatusInfo } from "./WorkspaceView";
+export type { WorkspaceStatusInfo } from "../../lib/types";
